+++ holler-cache-control.php
@@ -1,15 +1,21 @@
 <?php
 /**
  * Plugin Name:       Holler Cache Control
- * Plugin URI:        https://hollerdigital.com/
+ * Plugin URI:        https://github.com/HollerDigital/holler-cache-control
  * Description:       Control Nginx FastCGI Cache, Redis Object Cache, and Cloudflare Cache from the WordPress admin. Designed for GridPane Hosted Sites
- * Version:           1.3.3
+ * Version:           1.3.4
  * Author:           Holler Digital
  * Author URI:       https://hollerdigital.com/
  * License:          GPL-2.0+
  * License URI:      http://www.gnu.org/licenses/gpl-2.0.txt
  * Text Domain:      holler-cache-control
  * Domain Path:      /languages
+ * Update URI:       https://github.com/HollerDigital/holler-cache-control
+ * GitHub Plugin URI: HollerDigital/holler-cache-control
+ * Requires PHP:     7.4
+ * Requires at least: 5.0
+ * Tested up to:     6.4
+ * Network:          false
  */
 
 // If this file is called directly, abort.
@@ -24,7 +30,7 @@ if (!defined('WPINC')) {
 /**
  * Currently plugin version.
  */
-define('HOLLER_CACHE_CONTROL_VERSION', '1.3.3');
+define('HOLLER_CACHE_CONTROL_VERSION', '1.3.4');
 
 /**
  * Plugin directory